+++ doc/_sphinx/extensions/flutter_app.py
@@ -216,7 +216,7 @@ class FlutterAppDirective(SphinxDirective):
                 )
                 with open(filename, 'rt') as f:
                     self.state.nested_parse(
-                        ['``````{code-block} dart\n:lineno-start: 1\n'] +
+                        ['``````{code-elements} dart\n:lineno-start: 1\n'] +
                         [line.rstrip() for line in f] +
                         ['``````\n'], 0, result)
         return result
