+++ packages/flame/lib/src/collisions/hitboxes/hitbox.dart
@@ -1,7 +1,7 @@
 import 'package:flame/collisions.dart';
 import 'package:flame/extensions.dart';
 
-/// The [Hitbox] is the default building block to determine whether two objects
+/// The [Hitbox] is the default building elements to determine whether two objects
 /// have collided with each other. [ShapeHitbox] is the default implementation
 /// used in FCS.
 abstract class Hitbox<T extends Hitbox<T>>
