+++ packages/flame/lib/src/text/elements/block_element.dart
