+++ packages/flame/lib/src/text/elements/element.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-abstract class BackgroundElement {
+abstract class Element {
   void layout() {}
   void translate(double dx, double dy);
   void render(Canvas canvas);
